requirements.txt
ipdb
dotenv
pan-os-python
pan-python
ipaddress
rich
lxml
pywin32
requests